requirements.txt
pip~=24.2
Jinja2~=3.1.4
MarkupSafe~=3.0.2
Werkzeug~=3.1.1
click~=8.1.7
blinker~=1.8.2
itsdangerous~=2.2.0
mysqlclient~=2.2.5
Flask~=3.0.3
python-dotenv~=1.0.1
